--- PatriotRvWatchWidget/PatriotRvWatchWidget.intentdefinition
@@ -1,59 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<!DOCTYPE plist PUBLIC "-//Apple//DTD PLIST 1.0//EN" "http://www.apple.com/DTDs/PropertyList-1.0.dtd">
-<plist version="1.0">
-<dict>
-	<key>INEnums</key>
-	<array/>
-	<key>INIntentDefinitionModelVersion</key>
-	<string>1.2</string>
-	<key>INIntentDefinitionNamespace</key>
-	<string>88xZPY</string>
-	<key>INIntentDefinitionSystemVersion</key>
-	<string>20A294</string>
-	<key>INIntentDefinitionToolsBuildVersion</key>
-	<string>12A6144</string>
-	<key>INIntentDefinitionToolsVersion</key>
-	<string>12.0</string>
-	<key>INIntents</key>
-	<array>
-		<dict>
-			<key>INIntentCategory</key>
-			<string>information</string>
-			<key>INIntentDescriptionID</key>
-			<string>tVvJ9c</string>
-			<key>INIntentEligibleForWidgets</key>
-			<true/>
-			<key>INIntentIneligibleForSuggestions</key>
-			<true/>
-			<key>INIntentName</key>
-			<string>Configuration</string>
-			<key>INIntentResponse</key>
-			<dict>
-				<key>INIntentResponseCodes</key>
-				<array>
-					<dict>
-						<key>INIntentResponseCodeName</key>
-						<string>success</string>
-						<key>INIntentResponseCodeSuccess</key>
-						<true/>
-					</dict>
-					<dict>
-						<key>INIntentResponseCodeName</key>
-						<string>failure</string>
-					</dict>
-				</array>
-			</dict>
-			<key>INIntentTitle</key>
-			<string>Configuration</string>
-			<key>INIntentTitleID</key>
-			<string>gpCwrM</string>
-			<key>INIntentType</key>
-			<string>Custom</string>
-			<key>INIntentVerb</key>
-			<string>View</string>
-		</dict>
-	</array>
-	<key>INTypes</key>
-	<array/>
-</dict>
-</plist>
