+++ PatriotRvWidget/PatriotRvWidgetBundle.swift
@@ -13,6 +13,8 @@ struct PatriotRvWidgetBundle: WidgetBundle {
     //@WidgetBundleBuilder ?
     var body: some Widget {
         ChecklistWidget()
+        #if !os(watchOS)
         PowerWidgetLiveActivity()
+        #endif
     }
 }
